Model initial ventilator configuration
This change adds a number of interfaces, classes and enums to help us model the initial settings for our ventilator. This includes ventilation mode as well as the numerous set parameters for the ventilator. These should be used in our components as well as serial data parsing in the future. Lastly, a small example has been added to the HomeScreen to make use of this configuration and ensure it looks good