Add Alarms Banner
This change adds logic to generate notifications denoting alarm states. This logic will be used by an AlarmsBanner component which utilises the useReading hook. The logic checks the current active alarms, compares them to the new alarms value generated and takes action based on that - if no alarms, stop showing the banner, if new alarms, update the banner, if no change, then let the banner be.