Add alarm parsing
This change adds a couple of function in SerialParser to be able to parse the bytes containing alarm values. Since we have an array of alarms already, we just need to go through each of the byte values containing alarm information, find the value of each bit and append our alarms array with those which are set.

Close #21