Update ventilation mode parsing logic
The telemetry packet has changed and now send the ventilation mode as absolute values, present in 3 bits of the 29th byte. This change adds a new constant for ventilation modes and calculates the correct ventilation mode for the new packet. Lastly, a copy of the latest packet defintion has been included in the repository for reference. Lastly, the SerialParser file has been renamed to a .ts extension as it does not need to be .tsx