Separate processing of data packets (#139)
This change moves the processing logic of data packets into a separate component called `PacketProcessor`, which creates and fills a buffer with serial data before parsing and updating the reading context. To do this, it has been created to be flexible and take in a parse function as well as callback, to allow for easier mocking/dependency injection. To verify correctness of logic, a test has also been added.

#137