Fix build issues (#145)
Since the project has not been worked on for a while (6+ months), there were updates to the build systems which, if left unattended, would cause builds to fail. This was namely an NDK issue, but some other nuances in local builds were also complaining to do with gradle version. The recommended fixes have now been applied.
- [NDK Issues](https://stackoverflow.com/a/60474520)
- [Gradle versioning issues](https://stackoverflow.com/a/60850178)